requirements.txt
matplotlib==3.8.2
numpy==1.24.3
pandas==2.2.2
Pillow==10.1.0
Pillow==10.3.0
scipy==1.13.0
streamlit==1.33.0
